# annotrieve-api-specs.yaml
openapi: 3.1.0
info:
  version: "1.0.0"
  title: "Annotrieve API"
  description: "API for retrieving and managing eukaryotic genome annotations"
  contact:
    email: "emilio.righi@crg.eu"
servers:
  - url: https://genome.crg.es/annotrieve/api/v0
    description: Production server
...
paths:
...
  /assemblies:
    get:
      tags:
        - "assemblies"
      operationId: "getAssemblies"
      summary: "Get assemblies"
      parameters:
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/taxids"
        - $ref: "#/components/parameters/assembly_accessions"
        - $ref: "#/components/parameters/refseq_categories"
        - $ref: "#/components/parameters/assembly_levels"
        - $ref: "#/components/parameters/assembly_statuses"
        - $ref: "#/components/parameters/assembly_types"
        - $ref: "#/components/parameters/submitters"
        - $ref: "#/components/parameters/sort_by"
        - $ref: "#/components/parameters/sort_order"
      responses:
        "200":
          description: "Assemblies page"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAssembliesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - "assemblies"
      operationId: "postAssemblies"
      summary: "Get assemblies via POST"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssemblyQueryParams"
      responses:
        "200":
          description: "Assemblies page"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAssembliesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /assemblies/frequencies/{field}:
    get:
      tags:
        - "assemblies"
      operationId: "getAssemblyFrequencies"
      summary: "Get assembly field frequencies"
      parameters:
        - name: field
          in: path
          required: true
          description: "Field to aggregate. Allowed values: organism_name, submitter, source_database, assembly_level, refseq_category, assembly_status, assembly_type."
          schema:
            type: string
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/taxids"
        - $ref: "#/components/parameters/assembly_accessions"
        - $ref: "#/components/parameters/refseq_categories"
        - $ref: "#/components/parameters/assembly_levels"
        - $ref: "#/components/parameters/assembly_statuses"
        - $ref: "#/components/parameters/assembly_types"
        - $ref: "#/components/parameters/submitters"
      responses:
        "200":
          description: "Frequency counts"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrequencyCounts"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - "assemblies"
      operationId: "postAssemblyFrequencies"
      summary: "Get assembly field frequencies via POST"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssemblyQueryParams"
      responses:
        "200":
          description: "Frequency counts"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrequencyCounts"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /assemblies/{assembly_accession}:
    get:
      tags:
        - "assemblies"
      operationId: "getAssemblyById"
      summary: "Get assembly by accession"
      parameters:
        - name: assembly_accession
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Assembly details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assembly"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /assemblies/{assembly_accession}/chr-aliases:
    get:
      tags:
        - "assemblies"
      operationId: "getChromosomeAliases"
      summary: "Download chromosome alias table"
      description: "Streams a TSV file listing chromosome aliases for the assembly (chr_name, sequence_name, GenBank accession, RefSeq accession, UCSC name, aliases)."
      parameters:
        - name: assembly_accession
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "TSV stream"
          content:
            text/tab-separated-values:
              schema:
                type: string
                description: "Tab-separated alias table"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /assemblies/{assembly_accession}/chr_aliases:
    get:
      tags:
        - "assemblies"
      operationId: "getChromosomeAliasesDeprecated"
      summary: "Download chromosome alias table (deprecated)"
      deprecated: true
      description: "Deprecated. Use /assemblies/{assembly_accession}/chr-aliases instead. Streams a TSV file listing chromosome aliases for the assembly (chr_name, sequence_name, GenBank accession, RefSeq accession, UCSC name, aliases)."
      parameters:
        - name: assembly_accession
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "TSV stream"
          content:
            text/tab-separated-values:
              schema:
                type: string
                description: "Tab-separated alias table"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /assemblies/{assembly_accession}/assembled-molecules:
    get:
      tags:
        - "assemblies"
      operationId: "getAssembledMolecules"
      summary: "Download chromosomes.json for an assembly"
      description: "Streams chromosomes.json from LOCAL_ANNOTATIONS_DIR ({taxid}/{assembly_accession}/chromosomes.json), or from the paired assembly when the primary file is missing."
      parameters:
        - name: assembly_accession
          in: path
          required: true
          schema:
            type: string
        - name: offset
          in: query
          deprecated: true
          description: "Deprecated; ignored. Kept for backward compatibility."
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          deprecated: true
          description: "Deprecated; ignored. Kept for backward compatibility."
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        "200":
          description: "chromosomes.json (array of NCBI assembly report molecules)"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChromosomeReportEntry"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /assemblies/{assembly_accession}/assembled_molecules:
    get:
      tags:
        - "assemblies"
      operationId: "getAssembledMoleculesDeprecatedUnderscore"
      summary: "Download chromosomes.json for an assembly (deprecated)"
      deprecated: true
      description: "Deprecated. Use /assemblies/{assembly_accession}/assembled-molecules instead. Streams chromosomes.json from LOCAL_ANNOTATIONS_DIR."
      parameters:
        - name: assembly_accession
          in: path
          required: true
          schema:
            type: string
        - name: offset
          in: query
          deprecated: true
          description: "Deprecated; ignored. Kept for backward compatibility."
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          deprecated: true
          description: "Deprecated; ignored. Kept for backward compatibility."
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        "200":
          description: "chromosomes.json (array of NCBI assembly report molecules)"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChromosomeReportEntry"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /assemblies/{assembly_accession}/paired:
    get:
      tags:
        - "assemblies"
      operationId: "getPairedAssembly"
      summary: "Get the paired assembly"
      description: "Returns the paired RefSeq/GenBank assembly when available."
      parameters:
        - name: assembly_accession
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Paired assembly"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assembly"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    filter:
      name: "filter"
      in: "query"
      description: "Text search filter across multiple fields"
      schema:
        type: "string"
    limit:
      name: "limit"
      in: "query"
      description: "Maximum number of results to return"
      schema:
        type: "integer"
        minimum: 1
        maximum: 1000
        default: 20
    offset:
      name: "offset"
      in: "query"
      description: "Number of results to skip"
      schema:
        type: "integer"
        minimum: 0
        default: 0
    taxids:
      name: "taxids"
      in: "query"
      description: "Comma-separated list of taxonomy IDs"
      schema:
        type: "string"
      example: "9606,10090,7227"
...
    assembly_accessions:
      name: "assembly_accessions"
      in: "query"
      description: "Comma-separated list of assembly accessions"
      schema:
        type: "string"
      example: "GCA_000001405.29,GCF_000001405.40"
...
    refseq_categories:
      name: "refseq_categories"
      in: "query"
      description: "Comma-separated list of RefSeq categories"
      schema:
        type: "string"
    assembly_levels:
      name: "assembly_levels"
      in: "query"
      description: "Comma-separated list of assembly levels"
      schema:
        type: "string"
    assembly_statuses:
      name: "assembly_statuses"
      in: "query"
      description: "Comma-separated list of assembly statuses"
      schema:
        type: "string"
    assembly_types:
      name: "assembly_types"
      in: "query"
      description: "Comma-separated list of assembly types"
      schema:
        type: "string"
    submitters:
      name: "submitters"
      in: "query"
      description: "Comma-separated list of assembly submitter names"
      schema:
        type: "string"
...
    sort_by:
      name: "sort_by"
      in: "query"
      description: "Field to sort results by"
      schema:
        type: "string"
    sort_order:
      name: "sort_order"
      in: "query"
      description: "Sort order"
      schema:
        type: "string"
        enum: ["asc", "desc"]
...
  responses:
    BadRequest:
      description: "Bad request"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: "Resource not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: "Internal server error"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
      required:
        - detail
...
    AssemblyQueryParams:
      type: object
      properties:
        filter:
          type: string
        taxids:
          type: string
        bioproject_accessions:
          type: string
        assembly_accessions:
          type: string
        refseq_categories:
          type: string
        assembly_levels:
          type: string
        assembly_statuses:
          type: string
        assembly_types:
          type: string
        submitters:
          type: string
        sort_by:
          type: string
        sort_order:
          type: string
          enum: ["asc", "desc"]
        offset:
          type: integer
          default: 0
        limit:
          type: integer
          default: 20
...
    Assembly:
      type: object
      properties:
        assembly_accession:
          type: string
        paired_assembly_accession:
          type: string
        assembly_name:
          type: string
        source_database:
          type: string
        assembly_level:
          type: string
        assembly_status:
          type: string
        assembly_type:
          type: string
        refseq_category:
          type: string
        taxid:
          type: string
        organism_name:
          type: string
        taxon_lineage:
          type: array
          items:
            type: string
        release_date:
          type: string
          format: date-time
        submitter:
          type: string
        annotations_count:
          type: integer
        download_url:
          type: string
          format: uri
        assembly_stats:
          $ref: "#/components/schemas/AssemblyStats"
      required:
        - assembly_accession
        - assembly_name
        - organism_name
        - taxid
        - download_url
    AssemblyStats:
      type: object
      properties:
        total_number_of_chromosomes:
          type: integer
        total_sequence_length:
          type: string
        total_ungapped_length:
          type: string
        number_of_contigs:
          type: integer
        contig_n50:
          type: integer
        contig_l50:
          type: integer
        number_of_scaffolds:
          type: integer
        scaffold_n50:
          type: integer
        scaffold_l50:
          type: integer
        gaps_between_scaffolds_count:
          type: integer
        number_of_component_sequences:
          type: integer
        atgc_count:
          type: string
        gc_count:
          type: string
        gc_percent:
          type: integer
        genome_coverage:
          type: string
        number_of_organelles:
          type: integer
    ChromosomeReportEntry:
      type: object
      description: "One molecule row from chromosomes.json (NCBI assembly report)."
      properties:
        canonical_id:
          type: string
        chr_name:
          type: string
          description: "NCBI assembly report Assigned-Molecule (e.g. 1, 2, MT)"
        sequence_name:
          type: string
        ucsc_style_name:
          type: string
        genbank_accession:
          type: string
        refseq_accession:
          type: string
        sequence_role:
          type: string
          description: "NCBI assembly report Sequence-Role (e.g. assembled-molecule)"
        length:
          type: integer
...
    FrequencyCounts:
      type: object
      additionalProperties:
        type: integer
...
    PaginatedAssembliesResponse:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/Assembly"
...